+++ CHANGES.rst
@@ -5,6 +5,9 @@
 - Polish several key echelle methods
 - Modify create_linelist to default to vacuum
 - Update Xshooter, NIRES, and GNIRS
+- Refactor ProcessImages into ProcessImage, PypeItImage and CombinedImage
+- More precisely name the load() methods
+- Fix arc tilts bug
 
 0.10.1 (22 May 2019)
 --------------------